特技增加类型筛选

+++ biz/model/po/special_tech.go
@@ -3,6 +3,6 @@ package po
 type SpecialTech struct { // nolint:maligned
 	Id   int64  `gorm:"column:id" json:"id"`     // 主键ID
 	Name string `gorm:"column:name" json:"name"` // 名称
-	Type int    `gorm:"column:type" json:"type"` // 名称
+	Type int64  `gorm:"column:type" json:"type"` // 类型
 	Desc string `gorm:"column:desc" json:"desc"` // 描述
 }
